requirements.txt
flask==1.1.2
praw==7.2.0
python-dotenv==0.15.0